stash buffers using joined id
For reasons similar to the previous commit, stashing buffers
with ssl as key can lead to bad probe_reads at SSL_exit. Instead
a combination of both is needed. A new map that uses a concatenated
pid_tgid | ssl u128 value as key guarantees buffers will match for
a given id (count is not taken into account due to SSL_free not having
been called at this point).

In addition, an additional check has been added to poison_well in order
not to submit multiple pills for the same joined id.